fix(knowledge): address remaining code review feedback
- Handle empty traits in build_content to avoid trailing em dash
- Raise on unexpected popen3 calls in Regexp stub branch
- Add comment clarifying ESRCH rescue coverage in kill_process_group
- Extract exec_tmpdir helper to spec/support/exec_tmpdir.rb (DRY)
- Fix nil fallback: exec_tmpdir returns Dir.tmpdir when no exec dir found
